src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/host-context-class-change-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
Class change affecting a node in a distributed node's subtree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
